backend/requirements.txt
# FastAPI Backend for CodeFlow
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
